+++ nucliadb_node/src/replication/service.rs
@@ -21,6 +21,7 @@ use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
+use nucliadb_core::protos::{EmptyQuery, NodeMetadata};
 use nucliadb_core::tracing::{debug, warn};
 use nucliadb_core::NodeResult;
 use nucliadb_protos::replication;
@@ -274,5 +275,16 @@ impl replication::replication_service_server::ReplicationService for Replication
         });
 
         Ok(Response::new(ReceiverStream::new(receiver.1)))
+    }
+
+    async fn get_metadata(
+        &self,
+        _request: tonic::Request<EmptyQuery>,
+    ) -> Result<tonic::Response<NodeMetadata>, tonic::Status> {
+        let metadata = crate::node_metadata::NodeMetadata::new().await;
+        match metadata {
+            Ok(metadata) => Ok(tonic::Response::new(metadata.into())),
+            Err(error) => Err(tonic::Status::internal(error.to_string())),
+        }
     }
 }
