+++ nucliadb_protos/nodereader.proto
@@ -482,6 +482,4 @@ service NodeReader {
     // Shard Download
     rpc GetShardFiles(GetShardFilesRequest) returns (ShardFileList) {};
     rpc DownloadShardFile(DownloadShardFileRequest) returns (stream ShardFileChunk) {};
-
-    rpc GetMetadata(noderesources.EmptyQuery) returns (noderesources.NodeMetadata) {}
 }
