+++ nucliadb_protos/python/nucliadb_protos/nodereader_pb2_grpc.pyi
@@ -126,10 +126,6 @@ class NodeReaderStub:
         nucliadb_protos.nodereader_pb2.DownloadShardFileRequest,
         nucliadb_protos.nodereader_pb2.ShardFileChunk,
     ]
-    GetMetadata: grpc.UnaryUnaryMultiCallable[
-        nucliadb_protos.noderesources_pb2.EmptyQuery,
-        nucliadb_protos.noderesources_pb2.NodeMetadata,
-    ]
 
 class NodeReaderServicer(metaclass=abc.ABCMeta):
     """Implemented at nucliadb_object_storage"""
@@ -238,11 +234,5 @@ class NodeReaderServicer(metaclass=abc.ABCMeta):
         request: nucliadb_protos.nodereader_pb2.DownloadShardFileRequest,
         context: grpc.ServicerContext,
     ) -> collections.abc.Iterator[nucliadb_protos.nodereader_pb2.ShardFileChunk]: ...
-    @abc.abstractmethod
-    def GetMetadata(
-        self,
-        request: nucliadb_protos.noderesources_pb2.EmptyQuery,
-        context: grpc.ServicerContext,
-    ) -> nucliadb_protos.noderesources_pb2.NodeMetadata: ...
 
 def add_NodeReaderServicer_to_server(servicer: NodeReaderServicer, server: grpc.Server) -> None: ...
